feat: add support for default timeout, onNotify, and onError
Support for default timeout is added to allow clients to set a timeout
for request being made. Default timeout is not set by default. Timeout
is disabled by setting the timeout to undefined and enabled by setting
the timeout to a number to specify the timeout parameter in ms.

onNotify is used to receive notifications and onError is used to
receive errors that can't be attributed to a specific request.