feat: add support for custom generators
this change adds support for adding custom generators through
configuration, modules and .js files are able to be referred to
and hot loaded by the generator.

This will allow people to use the generator framework to generate
custom clients. The only requirement is that the generator modules
export an IComponentModule compatible object as the default export