fix: add package lock file and lock semi ver on a few packages
This is done because there was an issue with a dependency that
recently changed. The lock file is from a working version.

This is a temporary stop gap until we update gatsby and our deps
to later versions, which will result in a cleaner lock file and
a probable smaller footprint.