feat: add support for additional configuration to docs
This change allows for users to not need additional scripts
to override the gatsby config or docs, useful in automated
build situations and decouples docs from code more easily.

Additionally this changes adds support for static assets
to be picked up by gatsby useful for css and images and
are referenceable from the root path.

Add adds support for configuring the default index page
logo